Generify AmqAuthConnectedServiceBase to upper class.
Some minor methods are duplicated in child classes intentionally to
keep inheritance tree small. Rest is available in upper parent class
AmqpServiceBase.
