+++ services/device-registry/src/main/java/org/eclipse/hono/deviceregistry/SimpleDeviceRegistryServer.java
@@ -12,9 +12,19 @@
 
 package org.eclipse.hono.deviceregistry;
 
-import org.eclipse.hono.service.amqp.AmqpAuthConnectedServiceBase;
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.healthchecks.HealthCheckHandler;
+import io.vertx.ext.healthchecks.Status;
+import org.eclipse.hono.connection.ConnectionFactory;
+import org.eclipse.hono.service.amqp.AmqpServiceBase;
+import org.eclipse.hono.service.amqp.Endpoint;
+import org.eclipse.hono.service.auth.AuthenticationConstants;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
+
+import java.util.Objects;
 
 
 /**
@@ -25,10 +35,54 @@ import org.springframework.stereotype.Component;
  */
 @Component
 @Scope("prototype")
-public final class SimpleDeviceRegistryServer extends AmqpAuthConnectedServiceBase<DeviceRegistryConfigProperties> {
+public final class SimpleDeviceRegistryServer extends AmqpServiceBase<DeviceRegistryConfigProperties> {
+    protected ConnectionFactory authenticationService;
+
+    /**
+     * Sets the factory to use for creating an AMQP 1.0 connection to
+     * the Authentication service.
+     *
+     * @param factory The factory.
+     * @throws NullPointerException if factory is {@code null}.
+     */
+    @Autowired
+    @Qualifier(AuthenticationConstants.QUALIFIER_AUTHENTICATION)
+    public void setAuthenticationServiceConnectionFactory(final ConnectionFactory factory) {
+        authenticationService = Objects.requireNonNull(factory);
+    }
 
     @Override
     protected String getServiceName() {
         return "Hono-DeviceRegistry";
+    }
+
+    /**
+     * Registers this service's endpoints' readiness checks.
+     * <p>
+     * This invokes {@link Endpoint#registerReadinessChecks(HealthCheckHandler)} for all registered endpoints
+     * and it checks if the <em>Authentication Service</em> is connected.
+     *
+     * @param handler The health check handler to register the checks with.
+     */
+    @Override
+    public void registerReadinessChecks(final HealthCheckHandler handler) {
+        for (Endpoint ep : endpoints()) {
+            ep.registerReadinessChecks(handler);
+        }
+        handler.register("authentication-service-connection", status -> {
+            if (authenticationService == null) {
+                status.complete(Status.KO(new JsonObject().put("error", "no connection factory set for Authentication service")));
+            } else {
+                LOG.debug("checking connection to Authentication service");
+                authenticationService.connect(null, null, null, s -> {
+                    if (s.succeeded()) {
+                        s.result().close();
+                        status.complete(Status.OK());
+                    } else {
+                        status.complete(Status.KO(new JsonObject().put("error", "cannot connect to Authentication service")));
+                    }
+                });
+            }
+        });
     }
 }
